Réorganisation des fichiers sources | ajout de plusieurs Makefile
