+++ .gitignore
@@ -2,3 +2,4 @@ httpparseur
 .vscode
 get1
 objects/*.o
+objects/src/*.o
