+++ src/fonctions.c
