+++ src/parseur/headerConnection.c
@@ -1,0 +1,178 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include "../tree.h"
+#include "subroutineMultipleTerminal.h"
+#include "headerConnection.h"
+
+
+#define true 1
+#define false 0
+
+int Connection_header(int p, const char *req, node *pere)
+{
+    int len;
+    int now = p;
+
+    node* fils;
+
+    if(!(len = case_insensitive_string(now, req, fils = createFils(pere), "Connection"))) {
+        purgeNode(fils);
+        return false;
+    }
+    now += len;
+
+    if(!(len = case_insensitive_char(now, req, fils = createFrere(fils), ':'))) {
+        purgeNode(fils);
+        return false;
+    }
+    now += len;
+
+    if(!(len = OWS(now, req, fils = createFrere(fils)))) {
+        purgeFilsAndFrere(fils);
+    }
+    else {
+        fils = createFrere(fils);
+    }
+    now += len;
+
+    if(!(len = Connection(now, req, fils))) {
+        purgeNode(fils);
+        return false;
+    }
+    now += len;
+
+    if(!(len = OWS(now, req, fils = createFrere(fils)))) {
+        purgeNode(fils);
+    }
+    now += len;
+
+    putValueInNode(p, now-p, "Connection_header", pere);
+    return now-p;
+}
+
+int Connection(int p, const char *req, node *pere)
+{
+    int len;
+    int now = p;
+    node *fils;
+
+    int nbr = 0;
+    fils = createFils(pere);
+    while(len = case__AND__OWS(now, req, fils)) {
+        nbr++;
+        now += len;
+        fils = createFrere(fils);
+    }
+    purgeFilsAndFrere(fils);
+
+    if(!(len = connection_option(now, req, fils))) {
+        purgeNode(fils);
+        return false;
+    }
+    now += len;
+
+    nbr = 0;
+    while(len = OWS__AND__case__AND__optional(now, req, fils = createFrere(fils))) {
+        nbr++;
+        now += len;
+    }
+    purgeFilsAndFrere(fils);
+    purgeNode(fils);
+
+    putValueInNode(p, now-p, "Connection", pere);
+    return now-p;
+}
+
+int case__AND__OWS(int p, const char *req, node *pere)
+{
+    int len;
+    int now = p;
+
+    node* fils;
+
+    if(!(len = case_insensitive_char(now, req, fils = createFils(pere), ','))) {
+        purgeNode(fils);
+        return false;
+    }
+    now += len;
+
+    if(!(len = OWS(now, req, fils = createFrere(fils)))) {
+        purgeNode(fils);
+    }
+    now += len;
+
+    putValueInNode(p, now-p, "", pere);
+    return now-p;
+}
+
+int connection_option(int p, const char *req, node *pere)
+{
+    int len;
+    int now = p;
+
+    node* fils;
+
+    if(!(len = token(now, req, fils = createFils(pere)))) {
+        purgeNode(fils);
+        return false;
+    }
+    now += len;
+
+    putValueInNode(p, now-p, "connection_option", pere);
+    return now-p;
+}
+
+int OWS__AND__case__AND__optional(int p, const char *req, node *pere)
+{
+    int len;
+    int now = p;
+
+    node* fils;
+
+    if(!(len = OWS(now, req, fils = createFils(pere)))) {
+        purgeFilsAndFrere(fils);
+    }
+    else {
+        fils = createFrere(fils);
+    }
+    now += len;
+
+    if(!(len = case_insensitive_char(now, req, fils, ','))) {
+        purgeNode(fils);
+        return false;
+    }
+    now += len;
+
+    if(!(len = OWS__AND__connection_option(now, req, fils = createFrere(fils)))) {
+        purgeNode(fils);
+    }
+    now += len;
+
+    putValueInNode(p, now-p, "", pere);
+    return now-p;
+}
+
+int OWS__AND__connection_option(int p, const char *req, node *pere)
+{
+    int len;
+    int now = p;
+
+    node* fils;
+
+    if(!(len = OWS(now, req, fils = createFils(pere)))) {
+        purgeFilsAndFrere(fils);
+    }
+    else {
+        fils = createFrere(fils);
+    }
+    now += len;
+
+    if(!(len = connection_option(now, req, fils))) {
+        purgeNode(fils);
+        return false;
+    }
+    now += len;
+
+    putValueInNode(p, now-p, "", pere);
+    return now-p;
+}
