+++ src/parseur/headerContentLength.h
@@ -1,0 +1,10 @@
+#ifndef __HEADERCONTENTLENGTH_H__
+#define __HEADERCONTENTLENGTH_H__
+
+/*Prototype*/
+
+int Content_Length_Header(int p, const char *req, node *pere);
+
+    int Content_Length(int p, const char *req, node *pere);
+
+#endif
