+++ src/parseur/subroutineParseur.c
@@ -1,9 +1,12 @@
 #include <stdio.h>
 #include <stdlib.h>
-#include "tree.h"
+#include "../tree.h"
 #include "subroutineParseur.h"
 #include "subroutineMultipleTerminal.h"
 #include "subroutineFirstLine.h"
+
+#include "headerConnection.h"
+#include "headerContentLength.h"
 
 
 #define true 1
@@ -48,8 +51,6 @@ int HTTP_message(int p, const char *req, node *pere)
     putValueInNode(p, now-p, "HTTP_message", pere);
     return now-p;
 }
-
-
 
 int header_field__AND__CRLF(int p, const char *req, node *pere)
 {
@@ -140,239 +141,6 @@ int header_field(int p, const char *req, node *pere)
     return false;
 }
 
-int Connection_header(int p, const char *req, node *pere)
-{
-    int len;
-    int now = p;
-
-    node* fils;
-
-    if(!(len = case_insensitive_string(now, req, fils = createFils(pere), "Connection"))) {
-        purgeNode(fils);
-        return false;
-    }
-    now += len;
-
-    if(!(len = case_insensitive_char(now, req, fils = createFrere(fils), ':'))) {
-        purgeNode(fils);
-        return false;
-    }
-    now += len;
-
-    if(!(len = OWS(now, req, fils = createFrere(fils)))) {
-        purgeFilsAndFrere(fils);
-    }
-    else {
-        fils = createFrere(fils);
-    }
-    now += len;
-
-    if(!(len = Connection(now, req, fils))) {
-        purgeNode(fils);
-        return false;
-    }
-    now += len;
-
-    if(!(len = OWS(now, req, fils = createFrere(fils)))) {
-        purgeNode(fils);
-    }
-    now += len;
-
-    putValueInNode(p, now-p, "Connection_header", pere);
-    return now-p;
-}
-
-int Connection(int p, const char *req, node *pere)
-{
-    int len;
-    int now = p;
-    node *fils;
-
-    int nbr = 0;
-    fils = createFils(pere);
-    while(len = case__AND__OWS(now, req, fils)) {
-        nbr++;
-        now += len;
-        fils = createFrere(fils);
-    }
-    purgeFilsAndFrere(fils);
-
-    if(!(len = connection_option(now, req, fils))) {
-        purgeNode(fils);
-        return false;
-    }
-    now += len;
-
-    nbr = 0;
-    while(len = OWS__AND__case__AND__optional(now, req, fils = createFrere(fils))) {
-        nbr++;
-        now += len;
-    }
-    purgeFilsAndFrere(fils);
-    purgeNode(fils);
-
-    putValueInNode(p, now-p, "Connection", pere);
-    return now-p;
-}
-
-int case__AND__OWS(int p, const char *req, node *pere)
-{
-    int len;
-    int now = p;
-
-    node* fils;
-
-    if(!(len = case_insensitive_char(now, req, fils = createFils(pere), ','))) {
-        purgeNode(fils);
-        return false;
-    }
-    now += len;
-
-    if(!(len = OWS(now, req, fils = createFrere(fils)))) {
-        purgeNode(fils);
-    }
-    now += len;
-
-    putValueInNode(p, now-p, "", pere);
-    return now-p;
-}
-
-int connection_option(int p, const char *req, node *pere)
-{
-    int len;
-    int now = p;
-
-    node* fils;
-
-    if(!(len = token(now, req, fils = createFils(pere)))) {
-        purgeNode(fils);
-        return false;
-    }
-    now += len;
-
-    putValueInNode(p, now-p, "connection_option", pere);
-    return now-p;
-}
-
-int OWS__AND__case__AND__optional(int p, const char *req, node *pere)
-{
-    int len;
-    int now = p;
-
-    node* fils;
-
-    if(!(len = OWS(now, req, fils = createFils(pere)))) {
-        purgeFilsAndFrere(fils);
-    }
-    else {
-        fils = createFrere(fils);
-    }
-    now += len;
-
-    if(!(len = case_insensitive_char(now, req, fils, ','))) {
-        purgeNode(fils);
-        return false;
-    }
-    now += len;
-
-    if(!(len = OWS__AND__connection_option(now, req, fils = createFrere(fils)))) {
-        purgeNode(fils);
-    }
-    now += len;
-
-    putValueInNode(p, now-p, "", pere);
-    return now-p;
-}
-
-int OWS__AND__connection_option(int p, const char *req, node *pere)
-{
-    int len;
-    int now = p;
-
-    node* fils;
-
-    if(!(len = OWS(now, req, fils = createFils(pere)))) {
-        purgeFilsAndFrere(fils);
-    }
-    else {
-        fils = createFrere(fils);
-    }
-    now += len;
-
-    if(!(len = connection_option(now, req, fils))) {
-        purgeNode(fils);
-        return false;
-    }
-    now += len;
-
-    putValueInNode(p, now-p, "", pere);
-    return now-p;
-}
-
-int Content_Length_Header(int p, const char *req, node *pere)
-{
-    int len;
-    int now = p;
-
-    node* fils;
-
-    if(!(len = case_insensitive_string(now, req, fils = createFils(pere), "Content-Length"))) {
-        purgeNode(fils);
-        return false;
-    }
-    now += len;
-
-    if(!(len = case_insensitive_char(now, req, fils = createFrere(fils), ':'))) {
-        purgeNode(fils);
-        return false;
-    }
-    now += len;
-
-    if(!(len = OWS(now, req, fils = createFrere(fils)))) {
-        purgeFilsAndFrere(fils);
-    }
-    else {
-        fils = createFrere(fils);
-    }
-    now += len;
-
-    if(!(len = Content_Length(now, req, fils))) {
-        purgeNode(fils);
-        return false;
-    }
-    now += len;
-
-    if(!(len = OWS(now, req, fils = createFrere(fils)))) {
-        purgeNode(fils);
-    }
-    now += len;
-
-    putValueInNode(p, now-p, "Content_Length_Header", pere);
-    return now-p;
-}
-
-int Content_Length(int p, const char *req, node *pere)
-{
-    int len;
-    int now = p;
-    node *fils;
-
-    int nbr = 0;
-    fils = createFils(pere);
-    while(len = DIGIT(now, req, fils)) {
-        nbr++;
-        now += len;
-        fils = createFrere(fils);
-    }
-    purgeFilsAndFrere(fils);
-    purgeNode(fils);
-    if(nbr == 0) {
-        return false;
-    }
-    
-    putValueInNode(p, now-p, "Content_Length", pere);
-    return now-p;
-}
 
 int Content_Type_header(int p, const char *req, node *pere)
 {
